qcacmn: Implement a Mutex lock to safeguard the RNR database
When Thread 1 processes a scheduled scan (PNO) start request,
the host updates the rnr_channel_db. Simultaneously, if no
candidates are found in the scan database, Thread 2 initiates
a connect scan. To process the connect scan, the host flushes
the rnr_channel_db before updating it. During this time, there
is a possibility that Thread 1 might access the rnr_channel_db
while it is being flushed by Thread 2, leading to a potential
NULL pointer dereference and assertion failure.

The Fix is to protect the rnr_channel_db with a mutex lock
while it is being accessed by different threads.

Change-Id: I131eeb4c73af2553fa7c6c3d378ef6158813910d
CRs-Fixed: 3995483