qcacmn: Handle cfr stream init when STREAMFS not supported
In current implementation whenever SAP/Wi-Fi is turned on
API cfr_streamfs_init() is called, and a debugfs directory
is created using API qdf_streamfs_create_dir(). If WLAN_STREAMFS
is not enabled then the directory create fails due to
which cfr_streamfs_init() returns error resulting in excessive
error logging.

To handle this if WLAN_STREAMFS is not enabled, return status
of cfr_streamfs_init() as not supported.

CRs-Fixed: 4002063
Change-Id: I3330ecf6348127398d918128c86a32947dd3b9a6