icnss2: Complete sock wake completion if fw crashes
If fw crash happens, sock wake ack is not expected
and we need to fake complete this to prevent crash
in case fw boots up before fw down check. This
change also prevents increment in soc wake ref count
if ack is not received due to fw is down.

Change-Id: I3ce8d7ea2dcc0ead33080e00f86367e040e437c4
CRs-Fixed: 3999657