qcacld-3.0: Fix race between NB disconnect and Roam sync
Scenario:
1) Roam sync is being processed. Simultaneously, NB disconnect is
received, and SM moves to DISCONNECTING.
2) Scheduler thread process roam sync, and aborts roam since vdev
is DISCONNECTING.
3) As part of roam abort, RSO STOP is sent to firmware and RSO
SM moves to RSO_STOPPED.
4) Disconnect context is activated after serialization, and since
RSO SM is STOPPED, there is no need to wait for RSO_STOP resp,
and SM is moved to DEINIT and proceeds with cleanup.
5) This leads to out-of-sync issues between host and fw.

The expectation is that the host should wait until RSO STOP resp
and HO-Failure is received from firmware, before the cleanup is
done. To address this, do not send the RSO STOP from the roaming
context if disconnect request is not yet activated. Once activated,
send RSO STOP and wait for HO-FAIL from disconnect context.

Change-Id: Ia7154dad0293becf7296a7afa25626e91f911afd
CRs-Fixed: 4002178