cnss_genl: Add netink multicast group for opt_dp_logging
To implement a distinct logging mechanism for opt_dp logging,
introducing a new multicast group. This group is to ensure
that logs do not go to cnss diag and other userspace module.
Instead, only QHCP will listen to this group and
process the related netlink messages.

This change is to add netink multicast group
for opt_dp_logging.

Change-Id: Id0bf6d9fe29eaac7e3cf67250ad1075016172fdf
CRs-Fixed: 4007674