qcacmn: Allow SLO candidate if num links is more than max allowed
In current implementation, when candidate list is generated whether
SLO candidate is allowed is checked.

Consider:
1. If mlo_support_links is set to one, then only SLO candidate can be
considered.
2. And due to IOT OUI, SLO candidate is not allowed.

If only MLO candidate is present then it is not allowed and connection
cannot be made with AP.

To fix this, skip IOT check and add SLO candidate if num links is more
than mlo_support_links so STA can try lower number of links.

CRs-Fixed: 4003397
Change-Id: Ia7a6e4b347da904b9a35201113764bda14a6ddca