qcacmn: Skip force wake request/release when FW is down
Do not send force wake request/release when FW is down,
this is to make sure request and release are in
sync. Platform driver resets soc_wake_refcount
as part of ssr cleanup, so if host driver sends
soc wake release then it will mess up the ref
count. So skip sending force wake release
when FW is down.

Change-Id: Ib57baf75f4307cfa7ca38ba168c3e12436205457
CRs-Fixed: 4013246