qcacld-3.0: Fix invalid link validation during TDLS mgmt TX
Before transmitting the TDLS frame there is validation to check
if 2 of the links are on MCC. On single mac solution, we don't do
MCC on same links. So the validation is wrong for single mac
hardware.

Skip the STA links concurrency validation in TDLS management frame
TX path, for non-dbs hardware.

CRs-Fixed: 4012099
Change-Id: If7668c72231f848d13b17a6dbb61f3baec2aedcf