qcacld-3.0: Remove board ID info from wma service ready
Currently there is a provision to send the board ID by FW in
WMI service ready but this interface is no more used to send
this info and host is pritning this info as 0. This Board ID
information is sent using the QMI interface.

So remove the print in service ready indication for this board
ID.

CRs-Fixed: 3997932
Change-Id: Ic220efbf18b8adde773643b57c4f336fcc166d8d