cnss2: skip smmu fault processing when PCIe link is down
It will cause a TZ NOC read timeout error when processing
SMMU fault interrupt while the PCIe link is down. Below is
the call stack.
Although receiving an SMMU fault interrupt when the PCIe
link is down is not expected, to improve the robustness of
the code, we will skip SMMU fault processing in such cases.

[<fffffff054777abc>] readl_relaxed[cnss2]+0x38
[<fffffff05477d110>] cnss_mhi_read_reg[cnss2]+0x28
[<fffffff0545ea268>] mhi_get_mhi_state[mhi]+0x4c
[<fffffff0545f5524>] mhi_device_get_sync_atomic[mhi]+0xe0
[<fffffff05476b76c>] cnss_pci_smmu_fault_handler_irq[cnss2]+0x34
[<fffffff050223614>] arm_smmu_context_fault_irq[arm_smmu]+0xd8

Change-Id: I597db2ccf457d37b1398065f5cd1ed3fddc236bc
CRs-Fixed: 4019854