qcacmn: Don't flush conn BSSID during P2P scan with NL80211_SCAN_FLAG_FLUSH
Currently, when both STA and P2P are active, the driver
currently flushes all scan entries during a P2P scan
if the NL80211_SCAN_FLAG_FLUSH flag is set.
This results in the removal of the STA connected BSSID,
which causes issues during a link switch when the STA
cannot find the scan entry and gets disconnected.

Below is the scenario:
1. STA connected to ML AP with assoc_state set to
   SCAN_ENTRY_CON_STATE_ASSOC.
2. Driver does not flush the scan entry due to
   assoc_state.
3. P2P scan starts with NL80211_SCAN_FLAG_FLUSH set.
4. Driver flushes all scan entries and adds new ones,
   might be causing the connected BSSID entry to be
   aged out.
5. During link switch, STA cannot find the entry,
   leading to disconnection.

Fix is, don't flush the connected BSSID scan entry
even during P2P scan with the flush flag set.

CRs-Fixed: 4017723
Change-Id: I1da4f66e549f38c1296c02f9897a5381c850052f