qcacld-3.0: Move wakelock from peer to vdev
MLME private peer object create/destroy consumes more
time due to the wakelock create and destroy operations.
This in turn increases the latency in roam sync processing.
These wakelocks are acquired at the start of EAPOL and
released after EAPOL completion.

To reduce the latency of such operations, remove the
wakelock from peer level and move it to the vdev.
Maintain a counter to deactivate the runtime wakelock timer
upon the wakelock release of the last holder.

Change-Id: I386010e92c47a8a12ecbb66cc24547186be8ffe8
CRs-Fixed: 4014966