qcacmn: Update regulatory rules to wiphy during idle restart
Currently, during an idle restart, if the host receives a
channel avoid event before the regulatory pdev is open,
it fails to update the regulatory rules to wiphy as for
channel avoid event, updating reg rules to wiphy is not
needed. Additionally, it fails to process the channel
avoid event in this scenario.

To fix this, process the channel avoid event after the
regulatory pdev is open and then invoke the regulatory
regulatory northbound callback to update the regulatory
rules to wiphy.

Change-Id: Ic00cb28e91f7e8f5c90038b0077bf1d522f338f3
CRs-Fixed: 4011671