qcacmn: Reset link active state only during connect or roam
When link state switch event is received, the link state is
changed to active or inactive based on the received bitmap.
The active link is checked to allow TDLS. When probe response
frame is received from the AP, the link active flag is getting
reset and the subsequent TDLS discovery frame transmissions are
rejected.

Reset the link active state flag only during connect or roam

CRs-Fixed: 4038422
Change-Id: Id4cbb31eb8e6865676365ff721d8bb1860ae5c3c