qcacld-3.0: Correct the VHT caps in per-sta profile
The VHT CAP IE inside the ML IE is being populated with
the default caps. Therefore, some of the caps which are
specific to 5 GHz are being populated for the 2.4 GHz
partner as well.

To fix this, revise the vht caps in the VHT CAP IE inside
the per-sta profile using the link frequency.

Change-Id: Ifba4f6df6ae272219a74304e33a6fe4d32bf5a0c
CRs-Fixed: 4039413