qcacld-3.0: Add support for BW mitigation INI
The following INIs are introduced to support and manage the
BW mitigation feature:

1. ddr_bwm_sampling_time {Min: 10, Max: 100, Default: 100}
Configures the DBW mitigation sampling time (in ms) in the
firmware.

2. ddr_bw_mitigation_enable (default: disabled)
To control the BW mitigation feature via INI.

3. ddr_bwm_priority {Min: 1, Max: 10, Default: 1}
To control the priority of BW mitigation in the firmware.
The firmware considers this priority when applying the duty
cycle from multiple clients.

4. BWM throttle duty cycle for levels 0 to 5.

Change-Id: I5f7318f05d0ddf9b8083b2501ae4c84b410f40e0
CRs-Fixed: 4055747