qcacld-3.0: Acquire the SAP mutex lock if the HW mode changes
Currently, the host restarts SAP after a country change if the HW
mode changes. During the SAP restart, the host does not acquire the
SAP mutex lock for stopping BSS but only for starting BSS. This can
cause ref leak issues during vdev deletion in the following scenario:

1) As part of the SAP restart due to HW mode change, STOP BSS is
   in progress.
2) The host receives STOP_AP and, while handling it, takes the SAP
   mutex lock for stopping BSS.
3) After BSS stops, start BSS gets executed from the restart SAP path.
As a result, a reference leak is observed during SAP VDEV deletion
during the stop adapter.

To fix this issue, acquire the SAP mutex lock for SAP restart if the
HW mode changes as part of the country code change.

Change-Id: I30dff93b02f7b5e39c47d56781bf4a6ac2d1d770
CRs-Fixed: 4050193