qcacmn: Reduce the max value of scan mode duty cycle
When the host triggers a scan on the 6 GHz scan list, it depends on the
2 GHz or 5 GHz channels to decide whether to do a full scan or a scan
based on the RNR info. If there are no 2 GHz or 5 GHz channels in the
scan request and the duty cycle INI value is higher, it takes a long
time because multiple scans get triggered without any actual scan.

Reduce the maximum value of the 6 GHz scan mode duty cycle to 10

CRs-Fixed: 4038626
Change-Id: Iccbeff56818ffbb4baf184f7d33bfd26bf37eab3