qcacmn: Free old crypto keys before assigning new
Currently, while adding new key, driver finds the existing crypto
entry for the provided key and if crypto entry already exists for
that key, then it assigns new key to that entry. But driver does
not free the existing key memory. This can lead to memory leak
issue.

So, to fix this, free the existing key memory before assigning
new key in crypto entry.

Change-Id: Ia9821617b05e1e578201a6c6f2a36cc2978af7d0
CRs-Fixed: 4038503