qcacmn: Fix 11ax candidate BW score higher than 11be with puncture
When generating 11ax candidates for 11be AP with a puncture bitmap like
320-80, the puncture bitmap isn't cleared, leading to 11ax BW (40 MHz)
potentially being smaller than the punctured BW (80 MHz), which can cause
the full BW score to be miscalculated. This results in the 11ax candidate's
BW score being higher than that of 11be incorrectly, then connect as 11ax
mode.

To fix this, when generating 11ax candidates for 11be AP, we need to clear
the puncture bitmap, set puncture BW as 0 if no EHT CAP/OP in scan entry.

Change-Id: I95073a70c232d96a70ed2cbf67687f2242d128a6
CRs-Fixed: 4052856