touch: qts: get touch reset & intr base address from gpio number
The base address of the touch reset and interrupt GPIO is required
to access the GPIO registers in LE. It is hardcoded in the current
design. With this change, the touch interrupt and reset GPIOs are
parsed from the vendor touch and passed to QTS during client
registration. These GPIO numbers are used to get the base address
so that read/write operations on GPIO registers can be performed.

This helps to clear pending IRQ for the touch interrupt and toggle
the reset GPIO in case of TUI abort handling.

Change-Id: I396dce7c90c376f51eb931da2ef774a9f007ff9e
Signed-off-by: Anand Tarakh <quic_atarakh@quicinc.com>