msm: camera: isp: Handle flush case after init packet in flushed state
With stanby mode feature, 2 init packets are expected after the
flush call. In some corner cases, User space is triggering the
flush after sending the first init packet. This packet is received
in flush state. With current design this call is not handled. It results
in stale entries in the init packet request. When next init packet is
received, the BLs are getting corrupted with 2 requests.
This commit handles the flush call in flushed state and clears the
entries. This helps to avoid any stale entries in the request.

CRs-Fixed: 3914368
Change-Id: I6e0ecfa1814582cf5ef694d7ef44b36c3696efd0
Signed-off-by: Gaurav Jindal <quic_gjindal@quicinc.com>