qcacmn: Fix connected beacon interval filter command
In the current implementation of local packet capture, the connected
beacon interval filter is a subset of the connected beacon filter. This
means that the connected beacon interval filter only works if the
connected beacon filter is active. This deviates from the previous
implementation where the connected beacon interval was an
individual filter.

This change addresses the discrepancy between the current and
earlier implementations.

Change-Id: I380ee87e9b185fccc5fcbdd3ad73fb6014e30615
CRs-Fixed: 4064415