qcacld-3.0: Monitor mode init for non 11 BE
During monitor mode initialization, driver copies the link address
to peer mac address and register the DP peer with that address.
But for non 11 BE case, link address is zero. Hence DP peer
registration fails which leads to monitor mode initialization
failure.

So, to fix this, add 11 BE check. If device is 11 BE capable then
copy link address to peer address otherwise copy VDEV MAC address
to the peer address

Change-Id: I6c0901514788b18442e5370289eaa25859e40956
CRs-Fixed: 4047247