msm: camera: icp: Fix use-after-free issue in ICP driver
There are workqueues in ICP driver which are used to process
the msg received after submitting the commands to firmware via
HFI. These workqueue's execution is not in ICP driver's control.
During usecase closure, an abort and destroy command is sent to
firmware. The workqueue upon receipt of response from FW, marks
the completion event in ICP driver. If the workqueue's execution
is delayed by a very large time, it will lead to access of an
already freed resource and hence causing a use-after-free issue
in the driver.
To avoid this, ICP driver will also send the context id along
with context data which it will later use in workqueue to make
sure that the context data belongs to an active context.

CRs-Fixed: 4071529
Change-Id: Ic628b421f8254e8fc516a73d804e0b90e03ce413
Signed-off-by: Abhilash Kumar <quic_krabhi@quicinc.com>